refactor: '여정 등록' Trip에서 Itinerary로 도메인 변경
