+++ src/main/java/kr/co/fastcampus/travel/domain/itinerary/controller/ItineraryController.java
@@ -27,7 +27,6 @@ import org.springframework.web.bind.annotation.RestController;
 @RequiredArgsConstructor
 public class ItineraryController {
 
-    private final TripService tripService;
     private final ItineraryService itineraryService;
     private final ItineraryDtoMapper mapper;
 
@@ -38,7 +37,7 @@ public class ItineraryController {
         Principal principal
     ) {
         String memberEmail = principal.getName();
-        var response = tripService.addItineraries(
+        var response = itineraryService.addItineraries(
             request.tripId(), mapper.of(request), memberEmail
         );
         return ResponseBody.ok(mapper.of(response));
