+++ src/test/java/kr/co/fastcampus/travel/domain/itinerary/service/ItineraryServiceTest.java
@@ -18,7 +18,10 @@ import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
 
 import java.time.LocalDateTime;
+import java.util.List;
 import java.util.Optional;
+import java.util.stream.IntStream;
+import kr.co.fastcampus.travel.common.TravelTestUtils;
 import kr.co.fastcampus.travel.common.exception.EntityNotFoundException;
 import kr.co.fastcampus.travel.common.exception.InvalidDateSequenceException;
 import kr.co.fastcampus.travel.common.exception.MemberMismatchException;
@@ -26,6 +29,10 @@ import kr.co.fastcampus.travel.domain.itinerary.entity.Itinerary;
 import kr.co.fastcampus.travel.domain.itinerary.entity.Route;
 import kr.co.fastcampus.travel.domain.itinerary.entity.Transportation;
 import kr.co.fastcampus.travel.domain.itinerary.repository.ItineraryRepository;
+import kr.co.fastcampus.travel.domain.itinerary.service.dto.request.save.ItinerarySaveDto;
+import kr.co.fastcampus.travel.domain.itinerary.service.dto.request.save.LodgeSaveDto;
+import kr.co.fastcampus.travel.domain.itinerary.service.dto.request.save.RouteSaveDto;
+import kr.co.fastcampus.travel.domain.itinerary.service.dto.request.save.StaySaveDto;
 import kr.co.fastcampus.travel.domain.itinerary.service.dto.request.update.ItineraryUpdateDto;
 import kr.co.fastcampus.travel.domain.itinerary.service.dto.request.update.LodgeUpdateDto;
 import kr.co.fastcampus.travel.domain.itinerary.service.dto.request.update.RouteUpdateDto;
@@ -33,6 +40,8 @@ import kr.co.fastcampus.travel.domain.itinerary.service.dto.request.update.StayU
 import kr.co.fastcampus.travel.domain.itinerary.service.dto.response.ItineraryDto;
 import kr.co.fastcampus.travel.domain.member.entity.Member;
 import kr.co.fastcampus.travel.domain.trip.entity.Trip;
+import kr.co.fastcampus.travel.domain.trip.repository.TripRepository;
+import kr.co.fastcampus.travel.domain.trip.service.TripService;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
@@ -45,6 +54,8 @@ class ItineraryServiceTest {
 
     @Mock
     private ItineraryRepository itineraryRepository;
+    @Mock
+    private TripService tripService;
 
     @InjectMocks
     private ItineraryService itineraryService;
@@ -239,5 +250,139 @@ class ItineraryServiceTest {
         // when, then
         assertThatThrownBy(() -> itineraryService.editItinerary(id, member.getEmail(), request))
             .isInstanceOf(InvalidDateSequenceException.class);
+    }
+
+    @Test
+    @DisplayName("여정 복수 등록")
+    void addItineraries() {
+        // given
+        Member member = TravelTestUtils.createMember();
+        Trip trip = createTripWithMember(member);
+        List<ItinerarySaveDto> requests = IntStream.range(0, 3)
+            .mapToObj(i -> TravelTestUtils.createItinerarySaveDto())
+            .toList();
+
+        given(tripService.findById(trip.getId()))
+            .willReturn(trip);
+
+        //when
+        List<ItineraryDto> returnedItineraries = itineraryService.addItineraries(
+            trip.getId(), requests, member.getEmail()
+        );
+
+        //then
+        assertThat(returnedItineraries).isNotNull();
+        assertThat(returnedItineraries.size()).isEqualTo(3);
+    }
+
+    @Test
+    @DisplayName("여정 복수 등록 실패_없는 여행")
+    void addItineraries_fail() {
+        // given
+        Member member = TravelTestUtils.createMember();
+        Trip trip = createTripWithMember(member);
+        List<ItinerarySaveDto> requests = List.of();
+
+        given(tripService.findById(trip.getId()))
+            .willThrow(EntityNotFoundException.class);
+
+        //when
+        //then
+        assertThatThrownBy(() -> itineraryService.addItineraries(
+            trip.getId(), requests, member.getEmail()
+        )).isInstanceOf(EntityNotFoundException.class);
+    }
+
+    @Test
+    @DisplayName("여정 복수 등록 실패_작성자 불일치")
+    void addItineraries_fail_mismatchMember() {
+        // given
+        Member member = TravelTestUtils.createMember();
+        Trip trip = createTripWithMember(member);
+        List<ItinerarySaveDto> requests = List.of();
+
+        given(tripService.findById(trip.getId()))
+            .willReturn(trip);
+
+        //when
+        //then
+        assertThatThrownBy(() -> itineraryService.addItineraries(
+            trip.getId(), requests, "wrong_email"
+        )).isInstanceOf(MemberMismatchException.class);
+    }
+
+    @Test
+    @DisplayName("여정 추가 시 Lodge 종료일시가 시작일시보다 앞서면 예외")
+    void addItinerary_Lodge_InvalidDateSequence() {
+        // given
+        Member member = TravelTestUtils.createMember();
+        Trip trip = createTripWithMember(member);
+        given(tripService.findById(trip.getId()))
+            .willReturn(trip);
+
+        LodgeSaveDto lodgeSaveDto = new LodgeSaveDto("이름",
+            "주소",
+            LocalDateTime.of(2010, 1, 2, 12, 0),
+            LocalDateTime.of(2010, 1, 1, 12, 0)
+        );
+        ItinerarySaveDto saveDto = ItinerarySaveDto.builder()
+            .lodge(lodgeSaveDto)
+            .build();
+
+        // when, then
+        assertThatThrownBy(() -> itineraryService.addItineraries(
+            trip.getId(), List.of(saveDto), member.getEmail()
+        )).isInstanceOf(InvalidDateSequenceException.class);
+    }
+
+    @Test
+    @DisplayName("여정 추가 시 Route 종료일시가 시작일시보다 앞서면 예외")
+    void addItinerary_Route_InvalidDateSequence() {
+        // given
+        Member member = TravelTestUtils.createMember();
+        Trip trip = createTripWithMember(member);
+        given(tripService.findById(trip.getId()))
+            .willReturn(trip);
+
+        RouteSaveDto routeSaveDto = new RouteSaveDto(Transportation.BUS,
+            "출발지 이름",
+            "출발지 주소",
+            "도착지 이름",
+            "도착지 주소",
+            LocalDateTime.of(2010, 1, 2, 12, 0),
+            LocalDateTime.of(2010, 1, 1, 12, 0)
+        );
+        ItinerarySaveDto saveDto = ItinerarySaveDto.builder()
+            .route(routeSaveDto)
+            .build();
+
+        // when, then
+        assertThatThrownBy(() -> itineraryService.addItineraries(
+            trip.getId(), List.of(saveDto), member.getEmail()
+        )).isInstanceOf(InvalidDateSequenceException.class);
+    }
+
+    @Test
+    @DisplayName("여정 추가 시 Stay 종료일시가 시작일시보다 앞서면 예외")
+    void addItinerary_Stay_InvalidDateSequence() {
+        // given
+        Member member = TravelTestUtils.createMember();
+        Trip trip = createTripWithMember(member);
+        given(tripService.findById(trip.getId()))
+            .willReturn(trip);
+
+        StaySaveDto staySaveDto = new StaySaveDto("이름",
+            "주소",
+            LocalDateTime.of(2010, 1, 2, 12, 0),
+            LocalDateTime.of(2010, 1, 1, 12, 0)
+        );
+        ItinerarySaveDto saveDto = ItinerarySaveDto.builder()
+            .stay(staySaveDto)
+            .build();
+
+        // when, then
+        assertThatThrownBy(() -> itineraryService.addItineraries(
+            trip.getId(), List.of(saveDto), member.getEmail()
+        )).isInstanceOf(InvalidDateSequenceException.class);
     }
 }
